+++ databases/tests/test_batch.py
@@ -2,7 +2,6 @@ import pytest
 import prisma
 from prisma import Prisma
 
-from .utils import RawQueries
 from ..utils import DatabaseConfig
 
 
@@ -230,43 +229,6 @@ async def test_delete_many(client: Prisma) -> None:
         batcher.user.delete_many(where={'name': {'startswith': 'Robert'}})
 
     assert await client.user.count() == 0
-
-
-@pytest.mark.asyncio
-async def test_execute_raw(client: Prisma, raw_queries: RawQueries) -> None:
-    """execute_raw action can be used to execute raw SQL queries"""
-    post1 = await client.post.create(
-        {
-            'title': 'My first post!',
-            'published': False,
-        }
-    )
-    post2 = await client.post.create(
-        {
-            'title': 'My 2nd post.',
-            'published': False,
-        }
-    )
-
-    async with client.batch_() as batcher:
-        batcher.execute_raw(
-            raw_queries.update_unique_post_title,
-            post1.id,
-        )
-        batcher.execute_raw(
-            raw_queries.update_unique_post_new_title,
-            post2.id,
-        )
-
-    found = await client.post.find_unique(where={'id': post1.id})
-    assert found is not None
-    assert found.id == post1.id
-    assert found.title == 'My edited title'
-
-    found = await client.post.find_unique(where={'id': post2.id})
-    assert found is not None
-    assert found.id == post2.id
-    assert found.title == 'My new title'
 
 
 @pytest.mark.asyncio
