+++ databases/tests/test_update.py
@@ -1,7 +1,7 @@
 import pytest
 import prisma
 from prisma import Prisma
-from prisma.models import IntegerID, Unique2, User
+from prisma.models import IntegerID, User
 
 from lib.testing import async_fixture
 
@@ -213,27 +213,3 @@ async def test_update_id_field_atomic() -> None:
     )
     assert updated is not None
     assert updated.id == 501
-
-
-@pytest.mark.prisma
-@pytest.mark.asyncio
-async def test_update_unique_field() -> None:
-    """Setting a unique field"""
-    record = await Unique2.prisma().create(
-        data={
-            'name': 'Robert',
-            'surname': 'Craigie',
-        }
-    )
-
-    updated = await Unique2.prisma().update(
-        where={
-            'surname': record.surname,
-        },
-        data={
-            'surname': 'George',
-        },
-    )
-    assert updated is not None
-    assert updated.name == record.name
-    assert updated.surname == 'George'
